Moved ColonWidget to separate source files
