+++ Applications/Spire/Include/Spire/Ui/TimeOfDayInputWidget.hpp
@@ -30,36 +30,6 @@ namespace Spire {
       bool eventFilter(QObject* watched, QEvent* event) override;
 
     private:
-      class ColonWidget : public QWidget {
-        public:
-
-          explicit ColonWidget(QWidget* parent = nullptr)
-            : QWidget(parent),
-              m_border_color("#C8C8C8") {}
-
-          void set_style(const QString& color_hex) {
-            m_border_color = QColor(color_hex);
-            update();
-          }
-
-        protected:
-          void paintEvent(QPaintEvent* event) override {
-            auto painter = QPainter(this);
-            painter.fillRect(0, 0, width(), height(), Qt::white);
-            painter.fillRect(0, 0, width(), scale_height(1), m_border_color);
-            painter.fillRect(0, height() - scale_height(1), width(),
-              scale_height(1), m_border_color);
-            auto font = QFont("Roboto");
-            font.setPixelSize(scale_height(12));
-            painter.setFont(font);
-            painter.setPen(Qt::black);
-            painter.drawText(QPoint(0, scale_height(17)), ":");
-          }
-
-        private:
-          QColor m_border_color;
-      };
-
       mutable TimeChangeSignal m_time_signal;
       QLineEdit* m_hour_line_edit;
       ColonWidget* m_colon_widget;
